requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI integration
openai>=1.52.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Google Calendar integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0
pytz==2024.1

# Database - PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0

# Cache - Redis
redis==5.0.1
aioredis==2.0.1

# Monitoring & logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Async support
aiofiles==23.2.1
asyncio==3.4.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Health checks
healthcheck==1.3.3

# Rate limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6